34: Correct license headers
Task-Url: https://github.com/microprofile/microprofile-samples/issues/issues/34
