--- microprofile-sample-canonical/src/main/java/io/microprofile/sample/canonical/rest/TopCDsEndpoint.java
@@ -1,68 +1,0 @@
-/*
- * Copyright (C) 2016 Microprofile.io and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.microprofile.sample.canonical.rest;
-
-import io.microprofile.sample.canonical.utils.QLogger;
-
-import javax.enterprise.context.RequestScoped;
-import javax.inject.Inject;
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Random;
-import java.util.logging.Logger;
-
-@Path("/")
-@RequestScoped
-public class TopCDsEndpoint {
-
-    @Inject
-    @QLogger
-    private Logger logger;
-
-    @GET
-    @Produces(MediaType.APPLICATION_JSON)
-    public String getTopCDs() {
-
-        final JsonArrayBuilder array = Json.createArrayBuilder();
-        final List<Integer> randomCDs = getRandomNumbers();
-        for (final Integer randomCD : randomCDs) {
-            array.add(Json.createObjectBuilder().add("id", randomCD));
-        }
-        return array.build().toString();
-    }
-
-    private List<Integer> getRandomNumbers() {
-        final List<Integer> randomCDs = new ArrayList<>();
-        final Random r = new Random();
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-        randomCDs.add(r.nextInt(100) + 1101);
-
-        logger.info("Top CDs are " + randomCDs);
-
-        return randomCDs;
-    }
-}
