--- microprofile-sample-swagger/src/main/java/io/microprofile/sample/swagger/utils/ResourceProducer.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2016 Microprofile.io and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied.
- *
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.microprofile.sample.swagger.utils;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
-import javax.enterprise.inject.spi.InjectionPoint;
-import java.util.logging.Logger;
-
-@ApplicationScoped
-public class ResourceProducer {
-
-    @Produces
-    @QLogger
-    public Logger produceLogger(final InjectionPoint injectionPoint) {
-        return Logger.getLogger(injectionPoint.getMember().getDeclaringClass().getName());
-    }
-}
